Move, rename, BE comment handling files
