+++ lib/wordpress/comments/insertPostComment.js
@@ -1,5 +1,5 @@
-import mutationCommentToPost from '../comments/mutationCommentToPost'
 import {initializeWpApollo} from '../connector'
+import mutationInsertComment from './mutationInsertComment'
 
 /**
  * Add a comment to the given post. Follows established WordPress
@@ -54,7 +54,7 @@ export default async function insertPostComment(
   // Execute query.
   await apolloClient
     .mutate({
-      mutation: mutationCommentToPost,
+      mutation: mutationInsertComment,
       variables,
       context: {
         headers: {
