+++ lib/wordpress/comments/mutationInsertComment.js
@@ -2,8 +2,8 @@ import {gql} from '@apollo/client'
 import commentsFields from '../_partials/commentsFields'
 
 // Mutation: Add a comment to the given post
-const mutationCommentToPost = gql`
-  mutation ADD_COMMENT(
+const mutationInsertComment = gql`
+  mutation INSERT_COMMENT(
     $author: String
     $authorEmail: String
     $authorUrl: String
@@ -27,4 +27,4 @@ const mutationCommentToPost = gql`
   }
 `
 
-export default mutationCommentToPost
+export default mutationInsertComment
